Refactor code and add routes
