+++ models/place.js
@@ -5,6 +5,13 @@ var placeSchema = new mongoose.Schema({
   name: String,
   img: String,
   desc: String,
+  author: {
+    id: {
+      type: mongoose.Schema.Types.ObjectId,
+      ref: 'User'
+    },
+    username: String
+  },
   comments: [
     {
       type: mongoose.Schema.Types.ObjectId,
