+++ seeds.js
@@ -39,7 +39,7 @@ function seedDB() {
         console.log(err);
       }
       console.log('removed comments!');
-      //add a few campgrounds
+      // add a few campgrounds
       data.forEach(function(seed) {
         Place.create(seed, function(err, place) {
           if (err) {
