+++ views/comments/new.ejs
@@ -9,9 +9,6 @@
           <input class="form-control" type="text" name="comment[text]" placeholder="Comment">
         </div>
         <div class="form-group">
-          <input class="form-control" type="text" name="comment[author]" placeholder="Author">
-        </div>
-        <div class="form-group">
           <button class="btn btn-lg btn-primary btn-block">SUBMIT</button>
         </div>
       </form>
