+++ views/places/show.ejs
@@ -16,6 +16,9 @@
           <h4 class="pull-right">$9/visit</h4>
           <h4><a href=""><%= place.name %></a></h4>
           <p><%= place.desc %></p>
+          <p>
+            <em>Submitted By <%= place.author.username %></em>
+          </p>
         </div> <!-- closes caption-full -->
       </div> <!-- closes thumbnail -->
       <div class="well">
@@ -26,7 +29,7 @@
         <% place.comments.forEach(function(comment){ %>
           <div class="row">
             <div class="col-md-12">
-              <strong><%= comment.author %></strong>
+              <strong><%= comment.author.username %></strong>
               <span class="pull-right">10 days ago</span>
               <p>
                 <%= comment.text %>
